revert to r20 and add vs2008 project. nes only right now.

+++ trunk/src/nes/boards/emu2413.cpp
@@ -135,7 +135,7 @@ static void makeAdjustTable (OPLL * opll)
 
   opll->AR_ADJUST_TABLE[0] = (1 << EG_BITS);
   for (i = 1; i < 128; i++)
-    opll->AR_ADJUST_TABLE[i] = (e_uint16) ((double) (1 << EG_BITS) - 1 - (1 << EG_BITS) * log (i) / log (128));
+    opll->AR_ADJUST_TABLE[i] = (e_uint16) ((double) (1 << EG_BITS) - 1 - (1 << EG_BITS) * log ((double)i) / log (128.0));
 }
